rettet height og weight til doubles i unittest og controller

+++ LevelUp/Controller/UserController.cs
@@ -15,7 +15,7 @@ namespace Controller
 
         }
 
-        public User AddUserToDb(string username, string password, string name, int age, int weight, int height)
+        public User AddUserToDb(string username, string password, string name, int age, double weight, double height)
         {
             User demoUser = new User();
             int result = -1;
